+++ src/Factories/PinterestBot.php
@@ -6,6 +6,7 @@ use seregazhuk\PinterestBot\Bot;
 use seregazhuk\PinterestBot\Api\Request;
 use seregazhuk\PinterestBot\Api\CurlHttpClient;
 use seregazhuk\PinterestBot\Api\ProvidersContainer;
+use seregazhuk\PinterestBot\Api\Contracts\HttpClient;
 
 class PinterestBot
 {
@@ -19,14 +20,24 @@ class PinterestBot
      */
     public static function create($userAgent = "")
     {
-        $request = new Request(new CurlHttpClient());
-        if (!empty($userAgent)) {
-            $request->setUserAgent($userAgent);
-        }
-        
+        $request = new Request(self::getHttpClient($userAgent));
+
         $providersContainer = new ProvidersContainer($request);
 
         return new Bot($providersContainer);
+    }
+
+    /**
+     * @param string $userAgent
+     * @return HttpClient
+     */
+    protected static function getHttpClient($userAgent = "")
+    {
+        $httpClient = new CurlHttpClient();
+
+        if(!empty($userAgent)) $httpClient->setUserAgent($userAgent);
+
+        return $httpClient;
     }
 
     private function __construct()